pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["payne <antilevski@tut.by>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.16"
djangorestframework = "^3.14.0"
dynaconf = {extras = ["yaml"], version = "3.1.2"}
dj-database-url = "^1.0.0"
psycopg2-binary = "^2.9.5"
drf-spectacular = "^0.24.2"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
pytest-django = "^4.5.2"
pytest-cov = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
